api/requirements.txt
fastapi
uvicorn
pydantic
requests
httpx
openai
python-dotenv
jinja2
aiofiles